Remove ProcessManager interface.
